Fix http repository issues in pom.xml
Changes our dataone maven repo to use https and puts in an override mirror for for Restlet since the POM for the version of solr-core we use appears to include a link to the *http* restlet mirror. Maven as of 3.8.1  only supports https repository URLs and your entire dependency tree needs to respect that.